fix(component-rn): split pagination component
